refactor: extract rail auto init controller
